Move all files together
