--- client/shared/src/lib.rs
@@ -1,17 +1,0 @@
-#![feature(allocator_api)]
-#![feature(stdsimd)]
-#![feature(arbitrary_enum_discriminant)]
-
-pub mod alloc;
-pub mod double_buffer;
-pub mod error;
-pub mod mem;
-pub mod message_queue;
-pub mod sprite;
-pub mod state;
-pub mod wasm_log;
-
-pub use error::*;
-pub use mem::*;
-
-pub use wee_alloc;
