+++ wasm/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "rask-wasm"
+version = "0.1.0"
+authors = ["natrixaeria", "Dennis Kobert <dennis@kobert.dev>"]
+edition = "2018"
+description = "WebAssembly frontend of the rask project"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+wasm-bindgen = "=0.2.60"
+log = "0.4"
+js-sys = "=0.3.37"
+fern = "0.5.9"
+rand = "0.7"
+console_error_panic_hook = "0.1.6"
+const_env = "0.1"
+lazy_static = "1.4"
+
+[dependencies.rask-engine]
+version = "0.2.0"
+path = "../rask-engine"
+
+[dependencies.web-sys]
+version = "=0.3.36"
+features = [
+    "console",
+    "Window",
+    "OffscreenCanvas",
+    "WebGl2RenderingContext",
+    "DedicatedWorkerGlobalScope",
+    "WebGlShader",
+    "WebGlProgram",
+    "WebGlBuffer",
+    "WebGlVertexArrayObject",
+    "WebGlUniformLocation",
+    "WebGlFramebuffer",
+    "WebGlRenderbuffer",
+]
+
+[dependencies.wee_alloc]
+version = "=0.4.5"
+features = ["size_classes", "static_array_backend", "nightly", "extra_assertions"]
+git = "https://github.com/Ma27/wee_alloc"
+branch = "nightly"
