+++ wasm/Makefile.toml
@@ -28,7 +28,7 @@ script = [
 [tasks.crate-build]
 workspace = false
 condition = { env_set = [ "CARGO_TARGET_DIR" ] }
-env = { MEM_GEN = "${PWD}/gen/", RUSTFLAGS = "-Clink-arg=--stack-first -Clink-arg=--no-entry -Clink-arg=--import-memory -Clink-arg=--shared-memory  -Clink-arg=--threads -Clink-arg=--export=__wasm_init_memory -Clink-arg=--no-gc-sections -Clink-arg=--features=atomics,bulk-memory -Ctarget-feature=+atomics,+bulk-memory" }
+env = { MEM_GEN = "${PWD}/gen/", RUSTFLAGS = "-Clink-args=--stack-first -Clink-arg=--no-entry -Clink-arg=--import-memory -Clink-arg=--shared-memory  -Clink-arg=--threads -Clink-arg=--export=__wasm_init_memory -Clink-arg=--no-gc-sections -Clink-arg=--features=atomics,bulk-memory -Ctarget-feature=+atomics,+bulk-memory -Clink-arg=--verbose --emit obj -Z print-link-args" }
 script = [
 	"mkdir -p gen",
 	"touch shared/build.rs #triggers the build.rs to be rebuild every time",
