+++ wasm/src/alloc.rs
