+++ wasm/src/context.rs
