+++ wasm/src/double_buffer.rs
