+++ wasm/src/error.rs
