+++ wasm/src/game_context.rs
