+++ wasm/src/mem.rs
