+++ wasm/src/message_queue.rs
