+++ wasm/src/sprite.rs
