+++ wasm/src/state.rs
