+++ wasm/src/wasm_log.rs
